Move useLoadClassifications to classificationHooks Own file to reflect wider use than in JobBuilder
