+++ resources/assets/js/components/Application/BasicInfo/BasicInfoPage.tsx
@@ -23,7 +23,7 @@ import {
   useJobApplicationSteps,
   useTouchApplicationStep,
 } from "../../../hooks/applicationHooks";
-import { useLoadClassifications } from "../../../hooks/jobBuilderHooks";
+import { useLoadClassifications } from "../../../hooks/classificationHooks";
 
 interface BasicInfoPageProps {
   applicationId: number;
